Support for mv-id development and test environment
It is now possible to use a mv-id development or test environment as the
backend for login and servicecalls. A default is set in the
mvid-config.php file and empty values defaults to the production system.
The environment and context to use can also be set programatically
before including mvid-handler.php and an example of this can be found in
the index.php example file.